Added makefiles for cross-compilation to Win32 from linux systems. -Added Makefile-Nightly-* for compilation using mingw32. amd64 makefiles currently not working due to mysteriousness in the apt repo (*) -corrected cases in BLADPCED.RC, BLADES.RC, and BLSCENED.RC so they build properly on case-sensitive systems (*) -Fixed a typecasting error in dlogtool.cpp and a random comma error in tfileio.cpp (ScenEd) -Added *_private.rc files to repo to prevent build anarchy with windres.
git-svn-id: http://openexile.googlecode.com/svn/trunk@166 4ebdad44-0ea0-11de-aab3-ff745001d230

+++ Win32/Scenario Editor/dlogtool.cpp
@@ -573,7 +573,7 @@ short cd_kill_dialog(short dlog_num,short parent_message)
 		SetFocus(dlg_parent[which_dlg]);
 		SetWindowPos(dlg_parent[which_dlg],HWND_TOP,0,0,100,100,
 			SWP_NOMOVE | SWP_NOSIZE | SWP_NOREDRAW);
-		cd_set_edit_focus((short)dlg_parent[which_dlg]);
+		cd_set_edit_focus((size_t)dlg_parent[which_dlg]);
         }
 
 	if (parent_message > 0)
